added cmake

+++ homework-15/src/Utils/String.hpp
@@ -10,8 +10,8 @@ class Utils {
     // символы в этой строке могут занимать два байта в памяти. Поэтому, чтобы
     // определить длину строки в символах небюходимо от общего количества байт
     // отнять количество двухбайтовых символов.
-    int static stringLength(const std::string &s) {
-        const int twoBytesChars =
+    size_t static stringLength(const std::string &s) {
+        const size_t twoBytesChars =
             std::count_if(s.cbegin(), s.cend(), [](char c) -> bool {
                 // Многобайтовые символы начинаются с 10xxxxxx в двоичном
                 // формате.
